use zod as type source; move first helper to standalone file
